fix user panel/password edit

+++ src/components/userMenu.vue
@@ -3,15 +3,15 @@
     <mu-flat-button :label="email" ref="button" @click="toggle"></mu-flat-button>
     <mu-popover :trigger="trigger" :open="open" @close="handleClose">
       <mu-menu>
-        <mu-menu-item title="发送反馈" />
-        <mu-menu-item title="用户设置" />
-        <mu-menu-item title="登出" @click="dialog_open()" />
+        <mu-menu-item title="发送反馈" @click="dialog_open(2)" />
+        <mu-menu-item title="登出" @click="dialog_open(1)" />
       </mu-menu>
     </mu-popover>
-    <mu-dialog :open="dialog" title="登出系统" @close="dialog_close">
-      你确定要登出吗？
+    <mu-dialog :open="dialog" :title="dialog_title" @close="dialog_close">
+      <span v-if="dialog_page === 1">你确定要登出吗？</span>
+      <span v-if="dialog_page === 2">点击确定前往Github发表意见，请遵守<a href="https://gold.xitu.io/entry/5801d009570c35006c6ed781">《提问的艺术》</a></span>
       <mu-flat-button slot="actions" @click="dialog_close" primary label="取消"/>
-      <mu-flat-button slot="actions" primary @click="logout" label="确定"/>
+      <mu-flat-button slot="actions" primary @click="final" label="确定"/>
     </mu-dialog>
   </div>
 </template>
@@ -23,7 +23,9 @@ export default {
     return {
       open: false,
       trigger: null,
-      dialog: false
+      dialog: false,
+      dialog_title: '',
+      dialog_page: 1
     }
   },
   computed: {
@@ -41,17 +43,24 @@ export default {
     handleClose (e) {
       this.open = false
     },
-    dialog_open () {
+    dialog_open (page) {
+      this.dialog_page = page
+      if (page === 1) this.dialog_title = '登出系统'
+      else if (page === 2) this.dialog_title = '意见反馈'
       this.dialog = true
     },
     dialog_close () {
       this.dialog = false
     },
-    logout () {
+    final () {
       this.dialog_close()
-      this.$store.commit('logout')
-      window.localStorage.clear()
-      this.$router.push('/login')
+      if (this.dialog_page === 1) {
+        this.$store.commit('logout')
+        window.localStorage.clear()
+        this.$router.push('/login')
+      } else if (this.dialog_page === 2) {
+        location.href = 'https://github.com/noahziheng/freeiot/issues'
+      }
     }
   }
 }
